refactor: Updated /price-by-dates to a functional router
