+++ backend/api/products.py
@@ -2,6 +2,7 @@ from fastapi import APIRouter, HTTPException
 from typing import List, Optional
 from database import products_collection, price_by_date_collection
 from schemas import Product, Price_by_date
+from bson import ObjectId, errors
 
 router = APIRouter()
 
@@ -20,9 +21,12 @@ async def get_product(search: Optional[str] = None):
     return products
 
 
-@router.get("/price-by-dates", response_model=List[Price_by_date])
-async def read_price_by_dates():
-    price_by_dates = []
-    for price_by_date in price_by_date_collection.find():
-        price_by_dates.append(Price_by_date(**price_by_date))
-    return price_by_dates
+@router.get("/price-by-date", response_model=Price_by_date)
+async def get_price_by_date(search_id: Optional[str]):
+    if search_id is None:
+        raise HTTPException(status_code=404, detail="No search query provided")
+    price_by_date = price_by_date_collection.find_one({"_id": search_id})
+    if not price_by_date:
+        raise HTTPException(status_code=404, detail="No products found")
+    price_by_date = Price_by_date(**price_by_date)
+    return price_by_date
